[MISC][RL] Refactor form change handling to a hook
