+++ src/components/frontend-engine/frontend-engine.tsx
@@ -1,8 +1,7 @@
 import { yupResolver } from "@hookform/resolvers/yup";
 import cloneDeep from "lodash/cloneDeep";
 import isEmpty from "lodash/isEmpty";
-import isEqual from "lodash/isEqual";
-import { ReactElement, Ref, forwardRef, useCallback, useEffect, useImperativeHandle, useRef, useState } from "react";
+import { ReactElement, Ref, forwardRef, useCallback, useEffect, useImperativeHandle, useRef } from "react";
 import { FormProvider, useForm } from "react-hook-form";
 import useDeepCompareEffect, { useDeepCompareEffectNoCheck } from "use-deep-compare-effect";
 import {
@@ -19,8 +18,6 @@ import {
 	useFormSchema,
 	useFormValues,
 	useFrontendEngineForm,
-	usePrevious,
-	useValidationConfig,
 	useValidationSchema,
 } from "../../utils/hooks";
 import { Sections } from "../elements/sections";
@@ -31,21 +28,13 @@ import {
 	TFrontendEngineValues,
 	TWarningPayload,
 } from "./types";
+import { useFormChange } from "./use-form-change";
 
 const FrontendEngineInner = forwardRef<IFrontendEngineRef, IFrontendEngineProps>((props, ref) => {
 	// =============================================================================
 	// CONST, STATE, REFS
 	// =============================================================================
-	const {
-		data,
-		className = null,
-		components,
-		onChange,
-		onSubmit,
-		onSubmitError,
-		onValueChange,
-		wrapInForm = true,
-	} = props;
+	const { data, className = null, components, onSubmit, onSubmitError, wrapInForm = true } = props;
 	const {
 		className: dataClassName = null,
 		defaultValues,
@@ -66,7 +55,6 @@ const FrontendEngineInner = forwardRef<IFrontendEngineRef, IFrontendEngineProps>
 		rebuildValidationSchema,
 		yupId,
 	} = useValidationSchema();
-	const { formValidationConfig } = useValidationConfig();
 	const formMethods = useForm({
 		mode: validationMode,
 		reValidateMode: revalidationMode,
@@ -78,25 +66,10 @@ const FrontendEngineInner = forwardRef<IFrontendEngineRef, IFrontendEngineProps>
 		},
 	});
 	const { setFormSchema } = useFormSchema();
-	const {
-		reset,
-		watch,
-		handleSubmit: reactFormHookSubmit,
-		getValues,
-		setValue,
-		setError,
-		trigger,
-		formState,
-		clearErrors,
-	} = formMethods;
-	const { resetFields, setFields, setField, getFormValues, registeredFields } = useFormValues(formMethods);
+	const { reset, handleSubmit: reactFormHookSubmit, getValues, setValue, setError, trigger, formState } = formMethods;
+	const { resetFields, setFields, getFormValues, registeredFields } = useFormValues(formMethods);
 	const registeredFieldsRef = useRef(registeredFields); // using ref ensures the latest values can be retrieved in setErrors and setWarnings
-	const [oldFormValues, setOldFormValues] = useState<TFrontendEngineValues>({});
-	const previousFormValidationConfig = usePrevious(formValidationConfig);
-	const previousHardValidationSchema = usePrevious(hardValidationSchema);
-	const previousFormValues = useRef(undefined);
-	const previousFormValidity = useRef(undefined);
-	const [hasSchemaChange, setHasSchemaChange] = useState(false);
+	const { checkIsFormValid } = useFormChange(props, formMethods);
 
 	// =============================================================================
 	// HELPER FUNCTIONS
@@ -134,15 +107,6 @@ const FrontendEngineInner = forwardRef<IFrontendEngineRef, IFrontendEngineProps>
 		YupHelper.addCondition(type, name, fn, yupId, overwrite);
 		rebuildValidationSchema();
 	};
-
-	const checkIsFormValid = useCallback(() => {
-		try {
-			hardValidationSchema.validateSync(getValues());
-			return true;
-		} catch (error) {
-			return false;
-		}
-	}, [getValues, hardValidationSchema]);
 
 	const handleSubmit = useCallback((): void => {
 		onSubmit?.(getFormValues(undefined, stripUnknown));
@@ -194,113 +158,12 @@ const FrontendEngineInner = forwardRef<IFrontendEngineRef, IFrontendEngineProps>
 		addWarnings(newWarnings);
 	};
 
-	const handleValueChange = useRef(() => {
-		// noop
-	});
-	handleValueChange.current = () => {
-		const formValues = getFormValues(undefined, stripUnknown);
-		const formValidity = checkIsFormValid();
-		// attach / fire onValueChange event only when formValidationConfig has values
-		// otherwise isValid will be returned incorrectly as true
-		if (onValueChange && Object.keys(formValidationConfig || {}).length) {
-			// memoise to prevent unnecessary calls
-			if (previousFormValidity.current !== formValidity || !isEqual(previousFormValues.current, formValues)) {
-				onValueChange(formValues, formValidity);
-			}
-		}
-		previousFormValues.current = formValues;
-		previousFormValidity.current = formValidity;
-	};
-
 	// =============================================================================
 	// EFFECTS
 	// =============================================================================
 	useEffect(() => {
 		setFields(getValues());
 	}, []);
-
-	useEffect(() => {
-		const subscription = watch((value, { name }) => {
-			if (name) {
-				setField(name, value[name]);
-			} else {
-				setFields(value);
-			}
-		});
-		return () => subscription.unsubscribe();
-	}, []);
-
-	useEffect(() => {
-		// attach / fire onChange event only when formValidationConfig has values
-		// otherwise isValid will be returned incorrectly as true
-		if (onChange && Object.keys(formValidationConfig || {}).length) {
-			const subscription = watch(() => {
-				onChange(getFormValues(undefined, stripUnknown), checkIsFormValid());
-			});
-			onChange(getFormValues(undefined, stripUnknown), checkIsFormValid());
-
-			return () => subscription.unsubscribe();
-		}
-
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [checkIsFormValid, onChange, watch, formValidationConfig]);
-
-	useEffect(() => {
-		const subscription = watch(() => {
-			handleValueChange.current();
-		});
-
-		return () => subscription.unsubscribe();
-	}, [watch]);
-
-	useEffect(() => {
-		// when config changes due to overrides or conditional rendering, mark validity for re-evaluation
-		if (
-			previousFormValidationConfig &&
-			JSON.stringify(previousFormValidationConfig) !== JSON.stringify(formValidationConfig)
-		) {
-			setHasSchemaChange(true);
-		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [formValidationConfig]);
-
-	useEffect(() => {
-		// re-evaluate form validity when schema changes
-		if (hasSchemaChange && previousHardValidationSchema !== hardValidationSchema) {
-			handleValueChange.current();
-			setHasSchemaChange(false);
-		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [hasSchemaChange, hardValidationSchema]);
-
-	useEffect(() => {
-		const errors = formState.errors;
-
-		if (errors && !isEmpty(errors)) {
-			const subscription = watch((value) => {
-				const apiErrors = Object.fromEntries(
-					Object.entries(formState.errors).filter(([_, value]) => value.type === "api")
-				);
-				const hasApiErrors = apiErrors && !isEmpty(apiErrors);
-
-				if (hasApiErrors) {
-					Object.keys(apiErrors).forEach((key) => {
-						const oldValue = oldFormValues[key];
-						const updatedValue = value[key];
-
-						if (oldValue !== updatedValue) {
-							clearErrors(key);
-						}
-					});
-				}
-
-				setOldFormValues(value);
-			});
-
-			return () => subscription.unsubscribe();
-		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [formState, watch]);
 
 	useDeepCompareEffectNoCheck(() => {
 		reset(cloneDeep(defaultValues));
